Development

+++ pulse/broker/src/main/java/gov/ca/emsa/pulse/broker/dao/impl/PatientDAOImpl.java
@@ -31,8 +31,8 @@ public class PatientDAOImpl extends BaseDAOImpl implements PatientDAO {
 	@Override
 	public PatientDTO create(PatientDTO dto) {
 		PatientEntity patient = new PatientEntity();
-		patient.setFirstName(dto.getFirstName());
-		patient.setLastName(dto.getLastName());
+		patient.setGivenName(dto.getGivenName());
+		patient.setFamilyName(dto.getFamilyName());
 		patient.setDateOfBirth(dto.getDateOfBirth());
 		patient.setSsn(dto.getSsn());
 		patient.setGender(dto.getGender());
@@ -77,8 +77,8 @@ public class PatientDAOImpl extends BaseDAOImpl implements PatientDAO {
 	@Override
 	public PatientDTO update(PatientDTO dto) {
 		PatientEntity patient = this.getEntityById(dto.getId());
-		patient.setFirstName(dto.getFirstName());
-		patient.setLastName(dto.getLastName());
+		patient.setGivenName(dto.getGivenName());
+		patient.setFamilyName(dto.getFamilyName());
 		patient.setDateOfBirth(dto.getDateOfBirth());
 		patient.setSsn(dto.getSsn());
 		patient.setGender(dto.getGender());
@@ -124,8 +124,14 @@ public class PatientDAOImpl extends BaseDAOImpl implements PatientDAO {
 	@Override
 	public void delete(Long id) {
 		PatientEntity toDelete = getEntityById(id);
-		entityManager.remove(toDelete);
-		entityManager.flush();
+		if(toDelete != null) {
+			try {
+				entityManager.remove(toDelete);
+				entityManager.flush();
+			} catch(Exception ex) {
+				logger.error("Could not delete patient. Was it removed before the delete method was called?", ex);
+			}
+		}
 	}
